migraciones perfectamente configuradas, menú admin en proceso y el crud de semillero funcionando

+++ GestionSemilleros/GestionSemilleros/Models/Entidades/Actividad.cs
@@ -8,8 +8,8 @@ namespace GestionSemilleros.Models.Entidades
     public class Actividad
     {
         [Key]
-        public decimal IdActividad { get; set; }
-        public decimal IdFase { get; set; }
+        public int IdActividad { get; set; }
+        public int IdFase { get; set; }
         public string DuracionActividad { get; set; }
         public string NombreActividad { get; set; }
         public DateTime FechaEntregaActividad { get; set; }
