+++ GestionSemilleros/GestionSemilleros/Models/Entidades/Patrocinador.cs
@@ -8,7 +8,7 @@ namespace GestionSemilleros.Models.Entidades
     public class Patrocinador
     {
         [Key]
-        public decimal IdPatrocinador { get; set; }
+        public int IdPatrocinador { get; set; }
         public string NombrePatrocinador { get; set; }
         public string TipoPatrocinador { get; set; }
         public decimal TelefonoPatrocinador { get; set; }
