+++ GestionSemilleros/GestionSemilleros/Models/Entidades/Semillero.cs
@@ -8,7 +8,7 @@ namespace GestionSemilleros.Models.Entidades
     public class Semillero
     {
         [Key]
-        public decimal IdSemillero { get; set; }
+        public int IdSemillero { get; set; }
         public string NombreSemillero { get; set; }
         public string LineaSemillero { get; set; }
         public string EnfoqueSemillero { get; set; }
